mariadb slave_start支持重复启动

+++ role_mariadb_slave_config/files/docker-slave-auto-config.sh
@@ -2,16 +2,22 @@
 
 set -ex
 
+varContainerName=$varSlaveContainerName
+
+varDatabaseExistsCount=`mysql -uroot -p123456 -P3306 -h$varContainerName -e "show databases" | grep "$varMasterDatabaseName" | wc -l`
+if ! [[ $varDatabaseExistsCount -eq 0 ]]; then
+  echo "数据库同步容器之前已经初始化，不需要重新初始化，你可以忽略此提示信息"
+  exit
+fi
+
 # 解压/tmp/fullybackup-restore.tar.gz
 rm -f /tmp/*.sql
 (cd /tmp && tar -xvzf fullybackup-restore.tar.gz)
 (cd /tmp && mv *.sql fullybackup-restore.sql || mv .*.sql fullybackup-restore.sql)
 
-varContainerName=$varSlaveContainerName
-
-mysql -uroot -P3306 -h$varContainerName -e "create database $varMasterDatabaseName" && \
+mysql -uroot -p123456 -P3306 -h$varContainerName -e "create database $varMasterDatabaseName" && \
 echo "成功创建数据库$varMasterDatabaseName，正在使用全量备份还原数据库，可能需要等待一段时间。。。" && \
-mysql -uroot -P3306 -h$varContainerName $varMasterDatabaseName < /tmp/fullybackup-restore.sql && \
+mysql -uroot -p123456 -P3306 -h$varContainerName $varMasterDatabaseName < /tmp/fullybackup-restore.sql && \
 echo "成功还原数据库$varMasterDatabaseName"
 
 # 解析全量备份并获取master_log_file和master_log_pos
@@ -21,9 +27,9 @@ varMasterLogPos=`head -500 /tmp/fullybackup-restore.sql | grep "CHANGE MASTER TO
 rm -f /tmp/fullybackup-restore.sql
 
 if [[ "$varReplicationDelayInSeconds" == "" ]]; then
-  mysql -uroot -P3306 -h$varContainerName -e "change master to master_host='$varMasterIp',master_port=$varMasterPort,master_user='$varMasterReplicationUser',master_password='$varMasterReplicationUserPassword',master_log_file=$varMasterLogFile,master_log_pos=$varMasterLogPos" && echo "成功执行change master命令"
+  mysql -uroot -p123456 -P3306 -h$varContainerName -e "change master to master_host='$varMasterIp',master_port=$varMasterPort,master_user='$varMasterReplicationUser',master_password='$varMasterReplicationUserPassword',master_log_file=$varMasterLogFile,master_log_pos=$varMasterLogPos" && echo "成功执行change master命令"
 else
-  mysql -uroot -P3306 -h$varContainerName -e "change master to master_host='$varMasterIp',master_port=$varMasterPort,master_user='$varMasterReplicationUser',master_password='$varMasterReplicationUserPassword',master_log_file=$varMasterLogFile,master_log_pos=$varMasterLogPos,master_delay=$varReplicationDelayInSeconds" && echo "成功执行change master命令"
+  mysql -uroot -p123456 -P3306 -h$varContainerName -e "change master to master_host='$varMasterIp',master_port=$varMasterPort,master_user='$varMasterReplicationUser',master_password='$varMasterReplicationUserPassword',master_log_file=$varMasterLogFile,master_log_pos=$varMasterLogPos,master_delay=$varReplicationDelayInSeconds" && echo "成功执行change master命令"
 fi
 
-mysql -uroot -P3306 -h$varContainerName -e "start slave" && echo "成功启动slave复制进程"
+mysql -uroot -p123456 -P3306 -h$varContainerName -e "start slave" && echo "成功启动slave复制进程"
